refactor(frontend): extract full-page loading screen controller
